Add HDF5Writer to write combine output in hdf5 and support multiple channels, given by multiple input cards
